+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Relase Notes
 
+## @UNRELEASED_VERSION@
+
+### Validation
+
+Refactor validation and using external lib to show the diff if the table of content is outdated
+
 ## 0.4.4 (12.05.2021)
 
 ### Bugfix